+++ doc/source/changelog.rst
@@ -16,6 +16,7 @@ master: 2.4.0
 * Support for globally overloaded ordering operators
 * Special cases for __repr__/__str__ returning C++ stringy types
 * Fix lookup of templates of function with template args
+* Correct typing of int8_t/uint8_t enums
 
 
 2022-03-08: 2.3.0